extracted beer style map to another file
